+++ packages/deploy-config-sub-generator/src/headless/index.ts
@@ -2,11 +2,14 @@ import { join, resolve } from 'path';
 import { unlinkSync } from 'fs';
 import { DeploymentGenerator } from '@sap-ux/deploy-config-generator-shared';
 import { DeployTarget } from '@sap-ux/fiori-generator-shared';
-import { generatorNamespace, t } from '../utils';
+import { t, generatorNamespace, abapChoice, cfChoice } from '../utils';
 import type { CfDeployConfigOptions } from '@sap-ux/cf-deploy-config-sub-generator';
 import type { AbapDeployConfigOptions } from '@sap-ux/abap-deploy-config-sub-generator';
 import type { AppConfig } from '@sap-ux/fiori-app-sub-generator';
-import type { CFDeployConfig } from '@sap-ux/fiori-generator-shared';
+import type {
+    CFDeployConfig as CfHeadlessDeployConfig,
+    AbapDeployConfigOptions as AbapHeadlessDeployConfigOptions
+} from '@sap-ux/fiori-generator-shared';
 
 /**
  * Headless deployment generator
@@ -70,7 +73,8 @@ export default class extends DeploymentGenerator {
         const deployConfigOpts = this.transformExtConfig();
 
         if (deployConfigOpts && this.deployTarget) {
-            this.composeWith(generatorNamespace(this.genNamespace, this.deployTarget), {
+            const subGen = this.deployTarget === DeployTarget.CF ? cfChoice.name : abapChoice.name;
+            this.composeWith(generatorNamespace(this.genNamespace, subGen), {
                 arguments: this.args,
                 ...Object.assign(this.options, deployConfigOpts, {
                     telemetryData: additionalTelemetryData
@@ -88,23 +92,37 @@ export default class extends DeploymentGenerator {
         let options: CfDeployConfigOptions | AbapDeployConfigOptions | undefined;
 
         if (this.deployTarget === DeployTarget.CF) {
-            const cf = this.appConfig.deployConfig as CFDeployConfig;
+            const cf = this.appConfig.deployConfig as CfHeadlessDeployConfig;
             if (this.appConfig.project.targetFolder) {
                 options = {
                     projectRoot: this.appConfig.project.targetFolder,
                     destinationName: cf.destinationName,
                     destinationAuthType: cf.destinationAuthType,
                     addManagedAppRouter: cf.addToManagedAppRouter,
-                    launchDeployConfigAsSubGenerator: true,
+                    launchDeployConfigAsSubGenerator: true, // ensures prompting is skipped within cf sub gen
                     appRootPath: join(this.appConfig.project.targetFolder, this.appConfig.project.name),
                     addMTADestination: cf.addMTADestination,
                     lcapModeOnly: cf.lcapModeOnly,
                     cloudServiceName: cf.cloudServiceName
                 } as CfDeployConfigOptions;
             }
+        } else if (this.deployTarget === DeployTarget.ABAP) {
+            const abap = this.appConfig.deployConfig as AbapHeadlessDeployConfigOptions;
+            if (this.appConfig.project.targetFolder) {
+                options = {
+                    appRootPath: join(this.appConfig.project.targetFolder, this.appConfig.project.name),
+                    destination: abap?.destination,
+                    url: abap?.url ?? '', // URL may not be provided in BAS but will be determined in the generator
+                    client: abap?.client,
+                    scp: abap?.scp,
+                    ui5AbapRepo: abap?.ui5AbapRepo,
+                    description: abap?.description,
+                    package: abap?.package,
+                    transport: abap?.transport,
+                    launchDeployConfigAsSubGenerator: true // ensures prompting is skipping within abap sub gen
+                } satisfies AbapDeployConfigOptions;
+            }
         }
-        // ABAP support will be added in the future
-
         return options;
     }