+++ packages/flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @sap-ux/flp-config-sub-generator
 
+## 0.2.19
+
+### Patch Changes
+
+-   Updated dependencies [d6943aa]
+    -   @sap-ux/inquirer-common@0.7.8
+    -   @sap-ux/deploy-config-generator-shared@0.1.7
+    -   @sap-ux/flp-config-inquirer@0.3.15
+    -   @sap-ux/app-config-writer@0.6.8
+
+## 0.2.18
+
+### Patch Changes
+
+-   Updated dependencies [95a816d]
+    -   @sap-ux/fiori-generator-shared@0.12.6
+    -   @sap-ux/deploy-config-generator-shared@0.1.6
+    -   @sap-ux/flp-config-inquirer@0.3.14
+    -   @sap-ux/inquirer-common@0.7.7
+    -   @sap-ux/app-config-writer@0.6.8
+
+## 0.2.17
+
+### Patch Changes
+
+-   Updated dependencies [15ec5c4]
+    -   @sap-ux/fiori-generator-shared@0.12.5
+    -   @sap-ux/deploy-config-generator-shared@0.1.5
+    -   @sap-ux/flp-config-inquirer@0.3.13
+    -   @sap-ux/inquirer-common@0.7.6
+    -   @sap-ux/app-config-writer@0.6.8
+
+## 0.2.16
+
+### Patch Changes
+
+-   @sap-ux/app-config-writer@0.6.8
+-   @sap-ux/deploy-config-generator-shared@0.1.4
+-   @sap-ux/flp-config-inquirer@0.3.12
+
+## 0.2.15
+
+### Patch Changes
+
+-   Updated dependencies [58cea5c]
+    -   @sap-ux/app-config-writer@0.6.7
+
 ## 0.2.14
 
 ### Patch Changes